+++ tests/test_parcelbright.py
@@ -12,6 +12,8 @@ from datetime import datetime, timedelta
 import os
 import unittest
 
+from schematics.exceptions import ModelValidationError
+
 import parcelbright
 
 if not hasattr(unittest, 'skipUnless'):
@@ -20,64 +22,26 @@ if not hasattr(unittest, 'skipUnless'):
 
 class TestParcelBright(unittest.TestCase):
 
-    def test_parcel_entity(self):
-        p = parcelbright.Parcel(
-            width=10, height=10, length=10, weight=1
-        )
-        self.assertEqual(p.dict(), {
-            'width': 10, 'height': 10, 'length': 10, 'weight': 1,
-        })
-        self.assertEqual(
-            p.__repr__(),
-            '<Parcel [width=10, height=10, length=10, weight=1]>'
-        )
-
-    def test_parcel_entity_with_protected_fields(self):
-        p = parcelbright.Parcel(
-            width=10, height=10, length=10, weight=1
-        )
-        p._protected = True
-        p.__private = True
-        self.assertEqual(p.dict(), {
-            'width': 10, 'height': 10, 'length': 10, 'weight': 1,
-        })
-
-    def test_address_entity(self):
-        a = parcelbright.Address(
-            name='office', postcode='AAA AAA', town='London',
-            phone='12341234', country_code='GB', line1='line'
-        )
-        self.assertEqual(a.dict(), {
-            'name': 'office', 'postcode': 'AAA AAA', 'town': 'London',
-            'phone': '12341234', 'country_code': 'GB', 'line1': 'line'
-        })
-        self.assertEqual(
-            a.__repr__(),
-            '<Address [name=office, postcode=AAA AAA, town=London, line1=line, country_code=GB]>'  # NOQA
-        )
-
     def test_shipment_entity(self):
-        parcel = parcelbright.Parcel(
-            length=10, width=10, height=10, weight=1
-        )
-        from_address = parcelbright.Address(
-            name="office", postcode="NW1 0DU",
-            town="London", phone="07800000000",
-            line1="19 Mandela Street",
-            country_code="GB"
-        )
-        to_address = parcelbright.Address(
-            name="John Doe", postcode="E2 8RS",
-            town="London", phone="07411111111",
-            line1="19 Mandela Street",
-            country_code="GB"
-        )
-        shipment = parcelbright.Shipment(
-            customer_reference='123455667', estimated_value=100,
-            contents='books', pickup_date='2025-01-29',
-            parcel=parcel, from_address=from_address,
-            to_address=to_address
-        )
+        parcel = parcelbright.Parcel({
+            'width': 10, 'height': 10, 'length': 10, 'weight': 1
+        })
+        from_address = parcelbright.Address({
+            'name': 'office', 'postcode': 'NW1 0DU', 'town': 'London',
+            'phone': '07800000000', 'country_code': 'GB',
+            'line1': '19 Mandela Street'
+        })
+        to_address = parcelbright.Address({
+            'name': 'John Doe', 'postcode': 'E2 8RS', 'town': 'London',
+            'phone': '07411111111', 'country_code': 'GB',
+            'line1': '19 Mandela Street'
+        })
+        shipment = parcelbright.Shipment({
+            'customer_reference': '123455667', 'estimated_value': 100,
+            'contents': 'books', 'pickup_date': '2025-01-29',
+            'parcel': parcel, 'from_address': from_address,
+            'to_address': to_address
+        })
         self.assertEqual(
             shipment.__repr__(),
             '<Shipment [id=None, contents=books, state=unknown]>'
@@ -93,27 +57,28 @@ class TestParcelBright(unittest.TestCase):
     def test_rate(self):
         parcelbright.api_key = os.environ.get('PARCELBRIGHT_TEST_API_KEY')
         parcelbright.sandbox = True
-        parcel = parcelbright.Parcel(
-            length=10, width=10, height=10, weight=1
-        )
-        from_address = parcelbright.Address(
-            name="office", postcode="NW1 0DU",
-            town="London", phone="07800000000",
-            line1="19 Mandela Street",
-            country_code="GB"
-        )
-        to_address = parcelbright.Address(
-            name="John Doe", postcode="E2 8RS",
-            town="London", phone="07411111111",
-            line1="19 Mandela Street",
-            country_code="GB"
-        )
-        shipment = parcelbright.Shipment.create(
-            customer_reference='123455667', estimated_value=100,
-            contents='books', pickup_date='2025-01-29',
-            parcel=parcel, from_address=from_address,
-            to_address=to_address
-        )
+
+        parcel = parcelbright.Parcel({
+            'width': 10, 'height': 10, 'length': 10, 'weight': 1
+        })
+        from_address = parcelbright.Address({
+            'name': 'office', 'postcode': 'NW1 0DU', 'town': 'London',
+            'phone': '07800000000', 'country_code': 'GB',
+            'line1': '19 Mandela Street'
+        })
+        to_address = parcelbright.Address({
+            'name': 'John Doe', 'postcode': 'E2 8RS', 'town': 'London',
+            'phone': '07411111111', 'country_code': 'GB',
+            'line1': '19 Mandela Street'
+        })
+        shipment = parcelbright.Shipment({
+            'customer_reference': '123455667', 'estimated_value': 100,
+            'contents': 'books', 'pickup_date': '2025-01-29',
+            'parcel': parcel, 'from_address': from_address,
+            'to_address': to_address
+        })
+        shipment.create()
+
         self.assertTrue(isinstance(shipment.rates, list))
         self.assertEqual(shipment.state, 'incomplete')
 
@@ -140,27 +105,28 @@ class TestParcelBright(unittest.TestCase):
     def test_book_with_custom_pickup_date(self):
         parcelbright.api_key = os.environ.get('PARCELBRIGHT_TEST_API_KEY')
         parcelbright.sandbox = True
-        parcel = parcelbright.Parcel(
-            length=10, width=10, height=10, weight=1
-        )
-        from_address = parcelbright.Address(
-            name="office", postcode="NW1 0DU",
-            town="London", phone="07800000000",
-            line1="19 Mandela Street",
-            country_code="GB"
-        )
-        to_address = parcelbright.Address(
-            name="John Doe", postcode="E2 8RS",
-            town="London", phone="07411111111",
-            line1="19 Mandela Street",
-            country_code="GB"
-        )
-        shipment = parcelbright.Shipment.create(
-            customer_reference='123455667', estimated_value=100,
-            contents='books', pickup_date='2025-01-29',
-            parcel=parcel, from_address=from_address,
-            to_address=to_address
-        )
+
+        parcel = parcelbright.Parcel({
+            'width': 10, 'height': 10, 'length': 10, 'weight': 1
+        })
+        from_address = parcelbright.Address({
+            'name': 'office', 'postcode': 'NW1 0DU', 'town': 'London',
+            'phone': '07800000000', 'country_code': 'GB',
+            'line1': '19 Mandela Street'
+        })
+        to_address = parcelbright.Address({
+            'name': 'John Doe', 'postcode': 'E2 8RS', 'town': 'London',
+            'phone': '07411111111', 'country_code': 'GB',
+            'line1': '19 Mandela Street'
+        })
+        shipment = parcelbright.Shipment({
+            'customer_reference': '123455667', 'estimated_value': 100,
+            'contents': 'books', 'pickup_date': '2025-01-29',
+            'parcel': parcel, 'from_address': from_address,
+            'to_address': to_address
+        })
+        shipment.create()
+
         self.assertTrue(isinstance(shipment.rates, list))
         self.assertEqual(shipment.state, 'incomplete')
 
@@ -168,13 +134,77 @@ class TestParcelBright(unittest.TestCase):
             '%Y-%m-%d'
         )
         shipment.book(
-            shipment.rates[0]['code'], pickup_date=pickup_date
+            shipment.rates[0].code, pickup_date=pickup_date
         )
         self.assertEqual(shipment.state, 'completed')
-        self.assertEqual(shipment.pickup_date, pickup_date)
+        self.assertEqual(shipment.pickup_date.isoformat(), pickup_date)
 
         shipment.cancel()
         self.assertEqual(shipment.state, 'cancelled')
+
+    def test_address_validation(self):
+        # invalid name
+        with self.assertRaises(ModelValidationError):
+            parcelbright.Address({
+                'name': '', 'postcode': 'NW1 0DU',
+                'town': 'london', 'phone': '07800000000',
+                'line1': '12 Mandela Street',
+                'country_code': 'GB'
+            }).validate()
+
+        # invalid postcode
+        with self.assertRaises(ModelValidationError):
+            parcelbright.Address({
+                'name': 'name', 'postcode': '',
+                'town': 'london', 'phone': '07800000000',
+                'line1': '12 Mandela Street',
+                'country_code': 'GB'
+            }).validate()
+
+        # invalid town
+        with self.assertRaises(ModelValidationError):
+            parcelbright.Address({
+                'name': 'name', 'postcode': 'NW1 0DU',
+                'town': '', 'phone': '07800000000',
+                'line1': '12 Mandela Street',
+                'country_code': 'GB'
+            }).validate()
+
+        # invalid line1
+        with self.assertRaises(ModelValidationError):
+            parcelbright.Address({
+                'name': 'name', 'postcode': 'NW1 0DU',
+                'town': 'london', 'phone': '07800000000',
+                'line1': '',
+                'country_code': 'GB'
+            }).validate()
+
+        # invalid phone
+        with self.assertRaises(ModelValidationError):
+            parcelbright.Address({
+                'name': 'name', 'postcode': 'NW1 0DU',
+                'town': 'london', 'phone': '',
+                'line1': '12 Mandela Street',
+                'country_code': 'GB'
+            }).validate()
+
+        with self.assertRaises(ModelValidationError):
+            parcelbright.Address({
+                'name': 'name', 'postcode': 'NW1 0DU',
+                'town': 'london', 'phone': '078000000xx',
+                'line1': '12 Mandela Street',
+                'country_code': 'GB'
+            }).validate()
+
+        # invalid country_code
+        with self.assertRaises(ModelValidationError):
+            parcelbright.Address({
+                'name': 'name', 'postcode': 'NW1 0DU',
+                'town': 'london', 'phone': '07800000000',
+                'line1': '12 Mandela Street',
+                'country_code': ''
+            }).validate()
+
 
 if __name__ == '__main__':
     unittest.main()